Måste fixa shadowrays

+++ TNCG15/src/Ray.cpp
@@ -1,6 +1,6 @@
 #include "Ray.h"
 #include "Polygon.h"
 
-glm::vec3 Ray::getColor(glm::vec3 ray_col) {
+glm::dvec3 Ray::getColor(glm::dvec3 ray_col) {
 	return hit_surface->color * ray_col;
 }
